refactor: update file name to ReportZone
